+++ pytorch_lightning/loggers/wandb.py
@@ -25,7 +25,7 @@ class WandbLogger(LightningLoggerBase):
     --------
     .. code-block:: python
 
-        from pytorch_lightning.logging import WandbLogger
+        from pytorch_lightning.loggers import WandbLogger
         from pytorch_lightning import Trainer
 
         wandb_logger = WandbLogger()
